GCP_Agent_Starter_Pack/pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "winsights_agent"
version = "0.1.0"
description = "Flattened BigQuery agent (Winsights)"
authors = [{name="Samjad Azeez"}]
requires-python = ">=3.10"
dependencies = [
  "absl-py",
  "google-adk",
  "google-cloud-aiplatform",
  "python-dotenv",
  "pandas",
  "numpy",
  "matplotlib",
  "plotly",
]
